refactor : itemService 분리 및 코드 리팩토링
