+++ src/main/java/com/example/order/service/OrderService.java
@@ -27,62 +27,34 @@ public class OrderService {
 	private final ItemRepository itemRepository;
 	private final OrderRepository orderRepository;
 
-	public void start() {
-		String action = Input.main();
+	public boolean order() {
+		int orderNumber = Input.orderNumber();
 
-		if (action.equals("o")) {
-			Output.orderStart();
-			order();
-		} else if (action.equals("q") || action.equals("quit")) {
-			Output.orderExit();
-			System.exit(0);
-		} else {
-			throw new InvalidInputException();
+		if (orderNumber == 0) {
+			orderReceipt();
+			return false;
 		}
+
+		Item item = itemRepository.findByNumber(orderNumber)
+			.orElseThrow(InvalidInputException::new);
+
+		int orderQuantity = Input.orderQuantity();
+
+		if (orderQuantity == 0) {
+			orderReceipt();
+			return false;
+		}
+
+		Order order = new Order(orderQuantity);
+
+		item.addOrder(order);
+		item.decrease(orderQuantity);
+		return true;
 	}
 
-	public void order() {
-		int categoryNumber = Input.orderCategory();
-
-		List<Item> items = itemRepository.findAll();
-
-		Output.order(items, categoryNumber);
-
-		while (true) {
-			int orderNumber = Input.orderNumber();
-
-			if (orderNumber == 0) {
-				orderReceipt();
-				break;
-			}
-
-			Item item = itemRepository.findByNumber(orderNumber)
-				.orElseThrow(InvalidInputException::new);
-
-			int orderQuantity = Input.orderQuantity();
-
-			if (orderQuantity == 0) {
-				orderReceipt();
-				break;
-			}
-
-			Order order = new Order(orderQuantity);
-
-			item.addOrder(order);
-
-			stockDecrease(item, orderQuantity);
-		}
-		start();
-	}
-
-	public void orderReceipt() {
+	private void orderReceipt() {
 		List<Order> orders = orderRepository.findOrdersByStatus(OrderStatus.READY);
-
 		Output.orderReceipt(orders);
 		paymentService.payment(orders);
-	}
-
-	public void stockDecrease(Item item, Integer orderQuantity) {
-		item.decrease(orderQuantity);
 	}
 }
